INTERNAL: Upgraded Python2 scripts
Upgraded Python2 scripts to Python 3 using 2to3 tool.
This is to mitigrate security risks of Python2.7 deprecation.

+++ src/gallium/auxiliary/util/u_format_pack.py
@@ -44,7 +44,7 @@ from u_format_parse import *
 
 
 if sys.version_info < (3, 0):
-    integer_types = (int, long)
+    integer_types = (int, int)
 
 else:
     integer_types = (int, )
